+++ app/PhotonCms/Core/Entities/ModelRelation/RelationTypes/OneToOneExtended.php
@@ -112,7 +112,7 @@ class OneToOneExtended extends BaseRelation implements ModelRelationInterface, M
         }
         $content .= ";";
         FileContentHelper::addLinesAndIndent($content, 1, $amountOfIndent);
-        $content .= "\$table->foreign('{$this->sourceField}')->references('{$this->targetField}')->on('{$this->targetTable}')->onDelete('cascade');";
+        $content .= "\$table->foreign('{$this->sourceField}')->references('{$this->targetField}')->on('{$this->targetTable}')->onDelete('set null');";
         return $content;
     }